log service errors, update db table delta column type

+++ migrations/000001_initial.up.sql
@@ -2,7 +2,7 @@ CREATE TABLE metrics
 (
     id    VARCHAR NOT NULL,
     mtype VARCHAR NOT NULL,
-    delta INTEGER,
+    delta BIGINT,
     value DOUBLE PRECISION,
     hash  VARCHAR NOT NULL
 );
